third_party/blink/web_tests/svg/dynamic-updates/SVGFEDisplacementMapElement-dom-in-attr-expected.txt

SVG 1.1 dynamic update tests

Tests dynamic updates of the 'in' attribute of the SVGFEDisplacementMapElement object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS displacementMap.getAttribute('in') is "Map"
PASS displacementMap.getAttribute('in') is "Texture"
PASS successfullyParsed is true

TEST COMPLETE
